@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import { ucFirst } from 'change-case';
 import style from './traitConnection.css';
@@ -7,8 +8,8 @@ const TraitConnection = props => props.from && props.to ?
     (<div className={style.lineInvisible}/>);
 
 TraitConnection.propTypes = {
-    from: React.PropTypes.string,
-    to: React.PropTypes.string
+    from: PropTypes.string,
+    to: PropTypes.string
 };
 
 export default TraitConnection;
